boardend/requirements.txt
colorama==0.4.6
structlog==22.3.0
websocket-client==1.5.1
websockets==10.4
